% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ames}
\alias{ames}
\alias{stock}
\title{Ames House Price}
\format{
A data frame with columns:
\describe{
\item{Id}{House ID}
\item{GrLivArea}{Above grade (ground) living area square feet}
\item{LandSlope}{Slope of property}
\item{Neighborhood}{Physical locations within Ames city limits:
\describe{
\item{Blmngtn}{Bloomington Heights}
\item{Blueste}{Bluestem}
\item{BrDale}{Briardale}
\item{BrkSide}{Brookside}
\item{ClearCr}{Clear Creek}
\item{CollgCr}{College Creek}
\item{Crawfor}{Crawford}
\item{Edwards}{Edwards}
\item{Gilbert}{Gilbert}
\item{IDOTRR}{Iowa DOT and Rail Road}
\item{MeadowV}{Meadow Village}
\item{Mitchel}{Mitchell}
\item{Names}{North Ames}
\item{NoRidge}{Northridge}
\item{NPkVill}{Northpark Villa}
\item{NridgHt}{Northridge Heights}
\item{NWAmes}{Northwest Ames}
\item{OldTown}{Old Town}
\item{SWISU}{South & West of Iowa State University}
\item{Sawyer}{Sawyer}
\item{SawyerW}{Sawyer West}
\item{Somerst}{Somerset}
\item{StoneBr}{Stone Brook}
\item{Timber}{Timberland}
\item{Veenker}{Veenker}
}
}
\item{BldgType}{Building type:
\describe{
\item{1Fam}{Single-family Detached}
\item{2FmCon}{Two-family Conversion; originally built as one-family dwelling}
\item{Duplx}{Duplex}
\item{TwnhsE}{Townhouse End Unit}
\item{TwnhsI}{Townhouse Inside Unit}
}
}
\item{OverallCond}{Rates the overall condition of the house. 10 is Very Excellent, while 1 is Very Poor.}
\item{YearBuilt}{Original construction year}
\item{YrSold}{Year sold}
\item{SalePrice}{Sale price in dollars}
}
}
\usage{
data(ames)
}
\description{
House sales data for residential homes in Ames, Iowa. Data was compiled by Dean De Cock. Extracted from https://www.kaggle.com/c/house-prices-advanced-regression-techniques/overview
}
\keyword{datasets}
