% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{box}
\alias{box}
\title{Box office data}
\format{A data frame with 27499 rows and 11 variables}
\usage{
data(box)
}
\description{
The data contains weekly box office numbers as published on the website https://www.the-numbers.com/weekly-box-office-chart scraped on Aug 23 2018.
}
\details{
\itemize{
\item Rank current rank of the movie according to gross box office
\item Rank.last.Week last week's ranking of box office gross
\item Movie name of the movie
\item Distributor name of the Distributor
\item Gross weekly box office gross in US dollars
\item Change percent change in gross from last week
\item Thtrs. number of movie theaters the movie is being shown
\item Per.Thr. per theater gross
\item Total.Gross cumulative box office gross in US dollars
\item Date date of the publication of box office numbers
}
}
\keyword{datasets}
