% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget}
\alias{budget}
\title{Movie budget data}
\format{A data frame with 5524 rows and 5 variables}
\usage{
data(budget)
}
\description{
The data contains movie budgets and box office gross for the 5000 top grossing movies (as of Mar 6 2018) published
on https://www.the-numbers.com/movie/budgets/all/1
}
\details{
\itemize{
\item Release Date (date)
\item Movie name of the movie
\item Production Budget (dbl) budget in US dollars
\item Domestic Gross (dbl) in US dollars
\item Worldwide Gross (dbl) in US dollars
}
}
\keyword{datasets}
