% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Numbers of crimes by city}
\format{A data frame with 1207 rows and 14 variables:
\describe{
\item{City}{name of the city for which numbers are reported.}
\item{Population}{state population.}
\item{Violent.crime}{number of reported violent crimes.}
\item{Murder}{number of reported murders.}
\item{Rape}{number of reported rapes.  The definition of rape was updated in 2012. Numbers from 2013 reflect the new definition.}
\item{Robbery}{number of reported robberies.}
\item{Aggravated.assault}{number of reported aggravated assaults.}
\item{Property.crime}{total number of reported property crimes.}
\item{Burglary}{number of reported burglaries.}
\item{Larceny.theft}{number of reported larceny thefts.}
\item{Motor.vehicle.theft}{number of reported motor vehicle thefts.}
\item{Arson}{number of reported cases of arsons.}
\item{State}{name of the state for which numbers are reported.}
\item{Year}{year of the reporting.}
}}
\usage{
cities
}
\description{
A dataset containing the number of property and violent crimes across Iowa from 2006 to 2016.
The data was made available by the FBI in the Uniform Crime Reporting Statistics (UCR) at \url{https://www.ucrdatatool.gov/index.cfm}. The variables are as follows (more detail on the FBI website):
}
\keyword{datasets}
