% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_reg.R
\name{code_reg}
\alias{code_reg}
\title{Generates a region variable
\description{
\code{code_reg} generates a region variable using COW numeric ccode. 
}}
\usage{
code_reg(ccode, d)
}
\arguments{
\item{ccode}{The name of the variable in the data that contains COW ccode. 
Must be numeric.}

\item{d}{data.frame}

\item{...}{Numeric, complex, or logical vectors.}
}
\value{
the output will be a vector or the same length as ccode.
}
\description{
Generates a region variable
\description{
\code{code_reg} generates a region variable using COW numeric ccode. 
}
}
