% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_daily}
\alias{covid_daily}
\title{Daily Covid-19 cases in the US, March--November, 2020}
\format{
A data frame with 526323 rows and 9 variables:
\describe{
  \item{fips}{US county numeric identifier}
  \item{county}{county name}
  \item{date}{date}
  \item{state_name}{name of a US state}
  \item{population}{county population}
  \item{confirmed}{cumulative number of confirmed covid-19 cases}
  \item{new_confirmed}{new monthly count of confirmed covid-19 cases}
  \item{deaths}{cumulative number of confirmed covid-19 deaths}
  \item{new_deaths}{new monthly count of confirmed covid-19 deaths}
}
}
\source{
https://www.cnn.com/resources/coronavirus-information/
}
\usage{
covid_daily
}
\description{
A dataset on daily covid-19 cases in the US, between March and November, 2020, collected at the county level.
}
\keyword{datasets}
