% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defMat}
\alias{defMat}
\title{Defense alliances that formed in 1914}
\format{

}
\source{
109th Congress
}
\usage{
defMat
}
\description{
Defense alliances that formed in 1914
}
\keyword{datasets}
