% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{Earthquake data}
\format{A data frame with 22 variables}
\usage{
data(earthquakes)
}
\description{
The USGS monitors and reports earthquakes and earthquake-like events in almost real-time at https://www.usgs.gov/natural-hazards/earthquake-hazards.
For more information on the variables see https://earthquake.usgs.gov/data/comcat/data-eventterms.php
}
\details{
\itemize{
\item time date and time of the event in UTC
\item latitude geographic latitude
\item longitude geographic longitude
\item depth approximate depth of the event
\item mag magnitude of the event
\item magType method or algorithm used to calculate magnitude
\item nst total number of seismic stations used to determine earthquake location.
\item gap largest azimuthal gap between azimuthally adjacent stations (in degrees). In general, the smaller this number, the more reliable is the calculated horizontal position of the earthquake.
\item dmin Horizontal distance from the epicenter to the nearest station (in degrees). 1 degree is approximately 111.2 kilometers. In general, the smaller this number, the more reliable is the calculated depth of the earthquake.
\item rms Horizontal distance from the epicenter to the nearest station (in degrees). 1 degree is approximately 111.2 kilometers. In general, the smaller this number, the more reliable is the calculated depth of the earthquake.
\item net ID of a data contributor.
\item id unique identifier for the event
\item updated time when the event was most recently updated
\item place named geographic region near to the event
\item type type of seismic event.
\item horizontalError uncertainty of reported location of the event in kilometers.
\item depthError uncertainty of reported depth of the event in kilometers.
\item magError uncertainty of reported magnitude of the event.
\item magNst  total number of seismic stations used to calculate the magnitude for this earthquake.
\item locationSource network that originally authored the reported location of this event.
\item magSource network that originally authored the reported magnitude for this event.
}
}
\keyword{datasets}
