% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbi}
\alias{fbi}
\title{Numbers of crimes by state}
\format{
A data frame with 23672 rows and 7 variables:
\describe{
\item{State}{name of the state for which numbers are reported.}
\item{Abb}{two letter state abbreviation. }
\item{Year}{year of the reporting.}
\item{Population}{state population.}
\item{Type}{type of crime.}
\item{Count}{number of reported crimes.}
\item{Violent.crime}{logical value: is this crime against a person?.}
}
}
\usage{
fbi
}
\description{
A dataset containing the number of property and violent crimes across the United States from 1960 to 2017.
The data  from 1960 to 2014 was made available by the FBI in the Uniform Crime Reporting Statistics (UCR) at \url{https://www.ucrdatatool.gov/index.cfm}.
Since 2014 the data is only made available as part of Excel tables at \url{https://ucr.fbi.gov/crime-in-the-u.s/}.
The variables are as follows (more detail on the FBI website):
}
\keyword{datasets}
