% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbi.v2}
\alias{fbi.v2}
\title{Numbers of crimes by state and source}
\format{
A tibble with 26520 rows and 12 columns
\describe{
\item{state}{name of the state for which numbers are reported.}
\item{state_id}{id for each state. }
\item{state_abbr}{two letter state abbreviation. }
\item{year}{year of the reporting.}
\item{population}{state population.}
\item{type}{type of crime.}
\item{count}{number of reported offenses.}
\item{total_agency_count}{total number of crime-solving agencies in the state.}
\item{agency_submitting}{number of agencies who reported data.}
\item{population_covered}{percent of the state's population covered by reporting agencies.}
\item{source}{source of the estimate: SRS (standard reporting system) or
NIBRS (national incidence based reporting system)}
}
}
\usage{
fbi.v2
}
\description{
A dataset containing the state-wide counts of offenses for a selected number of crimes since 1979
as reported through the API for the FBI's Crime Data Explorer
at \url{https://cde.ucr.cjis.gov/LATEST/webapp/#/pages/docApi}.
Last updated: Sep 20 2023\if{html}{\out{<br>}}
The variables are as follows (more detail on the FBI website):
}
\examples{
# example code
library(tidyverse)

# compliance to report to NIBRS varies drastically by state
fbi.v2 \%>\% 
  ggplot(aes(x = year, y = agency_submitting/total_agency_count*100)) + 
    geom_point(aes(colour = source)) + 
    facet_wrap(~state, scales="free_y")

# population size is related to compliance
fbi.v2 \%>\% 
  filter(year==2021) \%>\%
  ggplot(aes(x = population, y = agency_submitting/total_agency_count*100)) + 
    geom_point() +
    geom_text(aes(label=state_abbr), 
              nudge_y = 3,
              data = fbi.v2 \%>\% 
                     filter(year==2021, 
                            agency_submitting/total_agency_count*100 < 50 | 
                            population > 2e+07) \%>\% unique())
    
# comparison of SRS and NIBRS counts in Iowa across all types of offenses
fbi.v2 \%>\% filter(state_abbr=="IA") \%>\%  
  ggplot(aes(x = year, y = count)) + 
    geom_point(aes(colour = source)) + 
    facet_wrap(~type, scales="free_y")

# comparison of SRS and NIBRS counts in New York across all types of offenses
fbi.v2 \%>\% filter(state_abbr=="NY") \%>\%  
  ggplot(aes(x = year, y = count)) + 
    geom_point(aes(colour = source)) + 
    facet_wrap(~type, scales="free_y")
}
\keyword{datasets}
