% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbiwide.v2}
\alias{fbiwide.v2}
\title{Numbers of crimes by state and source - wide format}
\format{
A tibble with 9139 rows and 17 columns
\describe{
\item{state}{name of the state for which numbers are reported.}
\item{state_id}{id for each state. }
\item{state_abbr}{two letter state abbreviation. }
\item{year}{year of the reporting.}
\item{population}{state population.}
\item{total_agency_count}{number of agencies in the state.}
\item{agency_submitting}{number of agencies involved in reporting to the system.}
\item{source}{source of the estimate: SRS or NIBRS.}
\item{homicide}{number of reported murders.}
\item{rape_legacy}{number of reported rapes before 2013. The definition of rape was updated in 2012 and reported afterwards as \code{rape_revised}.}
\item{rape_revised}{number of reported rapes using the definition of 2012.}
\item{robbery}{number of reported robberies.}
\item{aggravated_assaults}{number of reported aggravated assaults.}
\item{property_crime}{number of property crimes. This number should be the sum of the next four variables.}
\item{burglary}{number of reported burglaries.}
\item{larceny}{number of reported larceny thefts.}
\item{motor_vehicle_theft}{number of reported motor vehicle thefts.}
\item{arson}{number of reported incidents of arson.}
}
}
\usage{
fbiwide.v2
}
\description{
A dataset containing the number of property and violent crimes across the
United States as reported through the API for the FBI's Crime Data Explorer
at \url{https://cde.ucr.cjis.gov/LATEST/webapp/#/pages/docApi}.
Last updated: Sep 20 2023\if{html}{\out{<br>}}
The variables are as follows (more detail on the FBI website):
}
\keyword{datasets}
