% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLon2Sp.R
\name{LatLon2Sp}
\alias{LatLon2Sp}
\title{Translate latitude & longitude to the Cartesian coordinate on a unit sphere.}
\usage{
LatLon2Sp(lat, lon)
}
\value{
If \code{lat} and \code{lon} are scalars (length 1), then a vector for the Cartesian coordinate. If \code{lat} and \code{lon} have length n >= 1, then return a data frame with n rows and three columns names 'x', 'y', and 'z'
}
\description{
lat A vector containing the latitudes
lon A vector containing the corresponding longitudes
}
\examples{
# Get the Cartesian coordinates for Ames, IA
LatLon2Sp(42.0308, -93.6319)
}
