% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legnet}
\alias{legnet}
\title{Legislative cosponsorship data for 9 most ideological senators in the 109th Congress}
\format{

}
\source{
109th Congress
}
\usage{
legnet
}
\description{
Legislative cosponsorship data for 9 most ideological senators in the 109th Congress
}
\keyword{datasets}
