% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terr_attacks.wide}
\alias{terr_attacks.wide}
\title{Global Terrorist Attacks}
\format{
A data frame with 2,015 obs. of  16 variables:
\describe{
  \item{country}{target country name}
  \item{ccode}{target country COW numeric code}
  \item{cabb}{target country COW 3-letter code}
  \item{year}{year}
  \item{type}{attack type}
  \item{num_attacks}{number of attacks}
  \item{GDPpc}{target country GDP/capita (World Bank WDI)}
  \item{population}{target country population (COW)}
  \item{tradeofgdp}{target country trade as a percent of GDP (World Bank WDI)}
  \item{polity2}{target country Polity2 score (the Polity Project)}
  \item{Armed Assault}{number of terrorist attacks involving armed assault}
  \item{Bombing/Explosion}{number of terrorist attacks involving bombings/explosions}
  \item{Facility/Infrastructure Attack}{number of terrorist attacks targeting facilities/infrastructure}
  \item{Assassination}{number of terrorist attacks that were assassinations}
  \item{Hostage Taking (Kidnapping)}{number of terrorist attacks that involved kidnappings}
  \item{Unarmed Assault}{number of terrorist attacks that involved unarmed assault}
  \item{Hijacking}{number of terrorist attacks that involved hijackings}
  \item{Hostage Taking (Barricade Incident)}{number of terrorist attacks that involved barricading}
}
}
\source{
GTD dataset \url{https://www.start.umd.edu/data-tools/global-terrorism-database-gtd}
}
\usage{
terr_attacks.wide
}
\description{
A dataset of global terrorist attacks (based on GTD data) between 2001--2014. The variables are as follows (more detail on the GTD website).
}
\keyword{datasets}
