% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbi}
\alias{fbi}
\title{Numbers of crimes by state}
\format{
A data frame with 19476 rows and 8 variables:
\describe{
\item{state}{name of the state for which numbers are reported.}
\item{state_id}{id for each state. }
\item{state_abbr}{two letter state abbreviation. }
\item{year}{year of the reporting.}
\item{population}{state population.}
\item{type}{type of crime.}
\item{count}{number of reported crimes.}
\item{violent_crime}{logical value: is this crime against a person?.}
}
}
\usage{
fbi
}
\description{
A dataset containing the number of property and violent crimes across the
United States from about 1980 to 2020.
The data was acquired through the API for the FBI's Crime Data Explorer
at \url{https://cde.ucr.cjis.gov/LATEST/webapp/#/pages/explorer/crime/crime-trend}.
The variables are as follows (more detail on the FBI website):
}
\keyword{datasets}
