% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{happy}
\alias{happy}
\title{Data related to happiness from the general social survey.}
\format{
A data frame with 72390 rows and 12 variables
}
\usage{
data(happy)
}
\description{
The data is a small sample of variables related to happiness from the
general social survey (GSS). The GSS is a yearly cross-sectional survey of
Americans, run since 1972. We combine data from more than 30 samples to yield over 70 thousand
observations, and of the over 5,000 variables, we select some variables that are related to
happiness:
}
\details{
\itemize{
\item year. year of the response, 1972 to 2018.
\item age. age in years: 18--89 (89 stands for all 89 year olds and older).
\item degree. highest education: lt high school, high school, junior
college, bachelor, graduate.
\item finrela. how is your financial status compared to others: far below, below average, average, above average, far above.
\item happy. happiness: very happy, pretty happy, not too happy.
\item health. health: excellent, good, fair, poor.
\item marital. marital status:  married, never married, divorced,
widowed, separated.
\item sex. sex: female, male.
\item polviews. from extremely conservative to extremely liberal.
\item partyid. party identification: strong republican, not str republican, ind near rep, independent, ind near dem, not str democrat, strong democrat, other party.
\item wtssall. probability weight. applicable to years up to 2018
\item wtssnr. probability weight. applicable to years from 2004
}
}
\examples{
library(dplyr)
library(ggplot2)
happy \%>\% 
  filter(!is.na(happy), !is.na(sex)) \%>\%
  ggplot(aes(x = factor(year), fill = happy)) + 
    geom_bar(position = "fill") +
    facet_grid(sex~.) +
    scale_fill_brewer(palette="Greens")
}
\keyword{datasets}
