% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legnet}
\alias{legnet}
\title{Legislative cosponsorship data for 9 most ideological senators in the 109th Congress}
\format{
A data frame with 81 rows and 3 variables:
\describe{
  \item{name1}{name of the first senator}
  \item{name2}{name of the second senator}
  \item{cosponsor}{equals to 1 if the two senators cosponsored at least two bill}
}
}
\source{
109th Congress
}
\usage{
legnet
}
\description{
Legislative cosponsorship data for 9 most ideological senators in the 109th Congress
}
\keyword{datasets}
